requirements.txt
selenium==4.12.0
python-dotenv==1.0.0
webdriver-manager==4.10.0
yagmail==0.16.0
pytz==2024.1